+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - added `scrollToTopOfPage` to `OneBlinkFormBaseProps`
+- `pointAddressV3` element
 
 ## [8.12.2] - 2025-10-28